--- android/gitudy/domain/src/main/java/com/takseha/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.takseha.domain
-
-class MyClass {
-}
