+++ android/gitudy/presentation/build.gradle.kts
@@ -66,6 +66,6 @@ dependencies {
     implementation("com.google.dagger:hilt-android:2.50")
     kapt("com.google.dagger:hilt-android-compiler:2.50")
 
-    // androidx browser
-    implementation("androidx.browser:browser:1.5.0")
+    // androidx webkit
+    implementation("androidx.webkit:webkit:1.8.0")
 }
